pyproject.toml
[tool.poetry]
name = "python-cpi-api"
version = "0.1.0"
description = "Project to play with Python and GRPC"
authors = ["Wagner Tsuchiya <wagaun@gmail.com>"]
readme = "README.md"
packages = [{include = "python_cpi_api"}]

[tool.poetry.dependencies]
python = "^3.8.1"
grpcio-tools = "^1.54.2"
grpcio = "^1.54.2"
flake8 = "^6.0.0"
psycopg2-binary = "^2.9.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"

[tool.pytest.ini_options]
pythonpath = "src"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
